Create relevant parser files and minor refactoring
