+++ src/main/java/com/rocketden/tester/service/parsers/InputParser.java
@@ -1,0 +1,14 @@
+package com.rocketden.tester.service.parsers;
+
+import com.rocketden.tester.model.problem.ProblemTestCase;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class InputParser {
+
+    public List<Object> parseTestCase(ProblemTestCase testCase) {
+        return null;
+    }
+}
