+++ src/main/java/com/rocketden/tester/service/parsers/OutputParser.java
@@ -1,0 +1,14 @@
+package com.rocketden.tester.service.parsers;
+
+import com.rocketden.tester.dto.RunDto;
+import org.springframework.stereotype.Service;
+
+@Service
+public class OutputParser {
+
+    public RunDto parseOutput() {
+        // Requires discussion on how output will be stored and read in
+
+        return null;
+    }
+}
